app/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==2.5.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
SQLAlchemy==1.4.41
Werkzeug==2.3.7
Jinja2==3.1.2
alembic==1.12.0
python-dotenv==1.0.0
email-validator==2.0.0
WTForms==3.0.1
Flask-WTF==1.1.1
pytz==2023.3
python-dateutil==2.8.2
requests==2.31.0
gunicorn==21.2.0
pytest==7.4.2
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0
isort==5.12.0
mypy==1.5.1
coverage==7.3.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
six==1.16.0
urllib3==2.0.4
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
typing_extensions==4.7.1
